fenochem-backend/requirements.txt
fastapi
uvicorn
python-dotenv
byllm